Add the `ShellCalcJob` and `ShellParser` plugins
These `CalcJob` and `Parser` implementations make it possible to run a
shell command on a remote computer using the full power of the calcjob
engine.

The command to execute is represented by the `code` input which takes a
`Code` instance. It should contain the absolute path of the command that
is to be executed.

Any arguments that should be passed to the shell command can be
specified as a `List` node. Note that each argument here should be a
separate element in the list. This includes separating option flags and
their values, unless they are written as a single string joint by an
equal sign.

Certain shell commands take files as arguments. To have the job write
files to the working directory, specify them as `SinglefileData` nodes
for the values of a dictionary that is passed to the `files` input. The
keys of the input will be used as the filenames with which they are
written to the working directory. Note that since the keys are also used
for the link labels to attach the `SinglefileData` nodes to the process
nodes, they have to be valid link labels and so can only contain
alpha-numeric characters and non-consecutive underscores. If more
specific filenames are required, they can be defined using the
`filenames` input. Keys that overlap with those in `files` will use the
value of the `filenames` for the filename instead of the key.

The `arguments` input supports using string placeholders in the elements
which will be replaced with the filename of any files written to the
working directory, if the key corresponds in the `filenames` input
namespace.

The output written to the stdout and stderr file descriptors by the
command are captured and attached as `SinglefileData` output nodes. The
stderr is optional but the stdout is required. If no stdout was
retrieved the calculation is marked as failed.

Any other output files that need to be retrieved and attached as output
nodes can be specified through the `outputs` input. This takes a `List`
with relative filenames. The filenames can make use of the globbing
character `*` in case the exact output filenames that will be created
are not known in advance.