Add the `launch_shell_job` convenience wrapper function
This function makes it easy to launch a `ShellCalculation`. It takes
care of creating the necessary `Code` instance, if it doesn't already
exist and even creates and configures the localhost as `Computer` if not
already done and no explicit computer has been specified.

The goal is that a user can run a shell command with minimal setup code,
straight out of the box.