`ShellJob`: add validation for `outputs` input
The plugin will always redirect the stdout and stderr file descriptors
to files in the working directory named `stdout` and `stderr`. At the
end of the command, the exit code will be written to the file named
`status`. If a user were to invoke the `ShellJob` specifying any of
these names in the `outputs` input, the results may be unexpected as
files produced by the command may be overwritten.

Therefore, these three filenames are for now considered as reserved and
cannot be used as output filenames. A validator is added to the
`outputs` input port to catch this case.