`ShellJob`: add support for more `Data` types to `files`
So far the `files` input for the `ShellJob` made it possible to have the
job write input files to the working directory of the job. An additional
benefit was that, since the files were passed as `SinglefileData` nodes,
the provenance was automatically kept.

Besides files, however, there are use cases where other data types might
serve as input to a shell job. For example, a command could take command
line arguments that are a float, or integer or simple string. These
could in principle be communicated by the node equivalent in AiiDA, e.g,
a `Float`, `Int` or `Str` node, respectively. In contrast to the
`SinglefileData` nodes, their content would not be written to a file in
the working directory, but their content would be written as a string
and added to the command line arguments, replacing the corresponding
placeholder in the list of `arguments`.

Support is added for any `Data` type in addition to `SinglefileData`, as
long as it implements the `value` property which should return the node's
serialized into string form. A validator is added to the input port to
ensure that any nodes passed satisfy this requirement.

Since it now supports more than just "files", it makes sense to rename
the port from `files` to `nodes`. This might be slightly less intuitive
for beginning users as the majority of use-cases might simply need to
only pass `SinglefileData` nodes, but the increased consistency of
`nodes` is worth it in the end, since ultimately all users will anyway
have to understand the concept of a node at the very least.