`ShellJob`: Customize the `_build_process_label` method
This method returns the string that is set as the `process_label`
attribute on the nodes representing executions of the process class. By
default this is the name of the process class, i.e., `ShellJob` in this
case. This is not very informative as all `ShellJob` runs, regardless of
the shell command it ran, will get the same label.

Here we override the `_build_process_label` to include the full label of
the code, which in turn will include the label (typically the name of
the shell command) and the name of the computer on which is was run.
An example is `ShellJob<echo@localhost>`.