Add the `PickledData` data plugin
This `Data` implementation can be used to store (almost) any Python
object. It subclasses the `SinglefileData` and upon construction, the
Python object is pickled and the resulting bytes are written to the file
repository as a bytestream.

Instead of using the `pickle` module of the standard library as the
pickler, the `dill` library is used. This library is more versatile and
is capable of reliably pickling and unpickling a greater number of
Python objects, such as lambda's and functions.

The pickling and unpickling methods are configurable on the class. To
keep provenance, the module and name of the function used that should be
able to unpickle the pickled object, as well as the version of the
pickling package are stored in the attributes of the node.

The `load` method will attempt to load the Python object from the pickled
bytestream. It will call `get_unpickler` to load the unpickling method
which is loaded from the data stored in the attributes. If this fails, it
will instruct which package and what version should be installed to be
able to load the unpickler.