`launch_shell_job`: Change log level from warning to info
The `launch_shell_job` emits a log message on the warning level if no
explicit computer is specified and if the `Code` is created
automatically. However, given that warnings are always shown with the
default logging configuration of AiiDA and that most users will not
explicitly specify a code, as that is the whole point of this
convenience wrapper, they will get the warning message every time.

Instead, the log message is now emitted on the info level so it won't be
shown by default but can be made visible by changing the log level for
the AiiDA logger to `INFO`.