`ShellParser`: Fix output files with non-alphanumeric characters (#10)
Output files are attached using their filename as a link label. However,
the validation rules for link labels are a lot more restrictive compared
to filenames. The former can only contain alphanumeric characters and
non-consecutive underscores.

The `ShellParser` already replaced `.` characters, often found in
filenames used for the extension, into underscores, but these are not
the only illegal characters for link labels that can occur, such as
dashes etc. The parser is updated to replace any invalid character with
an underscore where consecutive underscores are merged into one.