`ShellJob`: Add the `ERROR_STDERR_NOT_EMPTY` exit code (#15)
Certain shell commands may fail to properly return a non-zero exit code
if the command failed. Currently, these shell jobs are marked as
finished successfully, which can be misleading. Often, the error will be
written to stderr which is a marker for the failure.

Here we add the `ERROR_STDERR_NOT_EMPTY` exit code that is returned by
the `ShellParser` if the exit status of the command is zero but the
stderr is not empty.