Add the `ShellCode` data plugin (#20)
The `ShellCode` data plugin is a subclass of the `InstalledCode` plugin
from `aiida-core`. The new plugin is used by the `launch_shell_job`
utility function whenever a code has to be created on the fly. This will
serve to be able to query for these kinds of codes and to distinguish
them from codes that are typically setup manually.