`launch_shell_job`: Move `arguments` as the second argument
When invoking `launch_shell_job` with positional arguments only, in most
cases users will naturally want to place the `arguments` input as the
second positional argument:

    launch_shell_job('some_command', '-p with -f arguments')

Currently, for these simple cases, the user is forced to use the keyword
argument notation since `arguments` does not come as second. This is
changed such that the most common case can be written as tersely as
possible.