`ShellJob`: Add `filename_stdin` input to redirect file through stdin
Certain codes require the input file to be passed in through the stdin
file descriptor. Currently there is no way to do this. One could try
to specify the `arguments` input as `['<', '{input_file}']` but the
redirection argument would be quoted and read as a literal argument
and so wouldn't work.

The `metadata.options.filename_stdin` input is added. When specified,
the filename is defined as the `stdin_name` attribute on the `CodeInfo`
instance prepared by `prepare_for_submission`. This will cause the
scheduler plugin to create the command execution line:

    executable < stdin_name

If `filename_stdin` is specified, the filename should be removed from
the `cmdline_arguments` attribute set on the `CodeInfo`, because
otherwise, the filename would be written twice on the command execution
line:

    executable stdin_name < stdin_name