`ShellJob`: Raise when `<` or `>` are specified in `arguments`
Stdout redirection is performed automatically by the `ShellJob` and
the `metadata.options.filename_stdin` input should be used to redirect a
particular file to stdin. Attempting to perform these redirections
"manually" through the `arguments` input will fail as the symbols will
be quoted and so interpreted as a literal command line argument.

To prevent users making this mistake, a validator is added for the
`arguments` input that validates that all elements are strings and that
the reserved symbols are not defined.