`ShellJob`: Automatically create parent directories in `filenames`
If a nested filepath was passed in the `filenames` input namespace, the
code would except with a `FileNotFoundError` if any of the parent
directories did not exist. The `ShellJob.write_single_file_data` method
is updated to always create the parent directories of a filepath before
attempting to write to it.