`ShellParser`: Add support to parse output directory as `FolderData`
So far, the `outputs` input could specify filepaths, which if it matched
a file, it would be wrapped as a `SinglefileData` and attached as an
output of the `ShellJob`. Here, support is added that if the output
filepath corresponds to a directory, it is wrapped as a `FolderData`
instead.

The `launch_shell_job` function is updated to return all output nodes in
the results dictionaries except for the `remote_folder` and the
`retrieved` outputs. This filtering used to be done on the type of the
output node since before only `SinglefileData` could be returned anyway
besides the `remote_folder` and `retrieved` outputs, but now that
`FolderData` nodes can also be parsed this filtering had to be updated.