`ShellJob`: Add the `additional_retrieve` option
This is required for the custom parser functionality. If the custom
parser needs to parse a file that is not retrieved by default, i.e, the
`stdout`, `stderr`, `status` or any of the files specified in the
`outputs` input, it won't be in the `dirpath` once the custom parser
gets called.

The typical way of ensuring a file is retrieved, by adding it to the
`outputs` list, is not a suitable solution, as this means the output
file is also attached as a `SinglefileData` node, which is likely not
what the user wants if they are already parsing it manually and
attaching its (partial) content in some other way.

For this use case, the `additional_retrieve` option is added, which,
when specified, is appended to the temporary retrieve list of the job.
This allows the user to specify which files to retrieve for the custom
parser.