DevOps: Fix the `daemon_client` fixture
The logic of the `DaemonClient` was changed in `aiida-core==2.3` such
that the timeout of the `stop_daemon` call is different from what is
used by the `is_daemon_running` property. This can lead to the stop call
going through just fine but `is_daemon_running` still returning `True`
straight after. This usually resolves after a while but this false
positive can cause the tests to fail when the session is closing. The
`stopped_daemon_client` was already fixed in `aiida-core` by adding a
manual grace period for `is_daemon_running` to start returning `False`.
Here the same fix is added for `daemon_client` by overriding it until it
has been fixed in `aiida-core`.