`launch_shell_job`: Accept string for the `arguments` argument (#36)
The `arguments` originally required a list of individual command line
arguments. This was because it is eventually passed to the `CalcJob`
which expects it in this form, such that each parameter can be properly
escaped for bash.

This form is not very user-friendly, however. Often users that will use
`aiida-shell` already have a workflow ready that they want to wrap with
AiiDA. Currently they have to manually split the command line arguments
into a list of individual args.

The `launch_shell_job` interface now accepts a `str` for `arguments` in
which case it will use `shlex.split` to split the string into a list of
individual arguments. The documentation is updated to exclusively use
this new form.