`launch_shell_job`: Fix bug in `submit=True` when used within work chain (#38)
When the `submit` argument of `launch_shell_job` is set to `True` the job
is submitted to the daemon using `aiida.engine.launch.submit`. However,
when this is executed within the context of a work chain, this raises an
error as `WorkChain.submit` should be used in this case instead of the
global function.

The `launch_shell_job` function is updated to check whether it is called
in the scope of a `WorkChain` by using `aiida.engine.Process.current()`.
If the case, the `WorkChain.submit` is used instead of the global launch
function.