`ShellJob`: Add support for `FolderData` in `nodes` input
The `nodes` input accepts any `Data` type that implements the `value`
property that returns a value that can be cast to a string. In addition,
a special case was added to add support for `SinglefileData` nodes which
would allow users to have files written to the working directory.

Here the logic is extended to also allow `FolderData` nodes. By default
all contents are written to the root of the working directory. The
`filenames` input namespace can be used to write the contents to a
subdirectory. For now it is not possible to write to nested directories
through this interface. However, in this case the user can create the
`FolderData` with the desired target hierarchy.