`ShellJob`: Do not copy contents of `nodes` to repository
The contents of all `SinglefileData` and `FolderData` nodes in the
`nodes` input namespace are written to the sandbox folder during the
`prepare_for_submission`. The contents of this folder are also copied to
the repository of the `CalcJobNode` representing the `ShellJob`
execution.

This is undesirable, however, since the content is essentially
duplicated in the data storage. The fact that all this content is
already indirectly attached to the `CalcJobNode` through the input links
of the relevant input node, it is not necessary to store it once again.

Therefore, all content written to the `folder` sandbox are added to the
`provenance_exclude_list` which ensures that the engine does not copy
the contents to the `CalcJobNode` repository.