Add the `EntryPointData` data plugin
This data plugin wraps an existing `EntryPoint`. It can be constructed
from an `EntryPoint` instance or the entry point name and group. In
addition to storing the attributes of the `EntryPoint` itself in the
node's attributes (value, name, group, module, attr, and extras) the
plugin attempts to get the version of the package that provides the
entry point. The `load` method can be used on an `EntryPointData`
instance to reconstruct the `EntryPoint` and load it.