Refactor: Move module `engine.launchers.shell_job` to `launch`
The nested structure of the `engine.launchers` module stems from the
very first conception of `aiida-shell` when there was an implementation
based on process functions, in addition to the `CalcJob` implementation.
each of which would get its own launcher.

Now that only the `ShellJob` is left and the only one that will be
necessary, the nesting is no longer needed.