`ShellJob`: Add support for `RemoteData` nodes (#48)
The `ShellJob` plugin is updated to accept `RemoteData` nodes in the
`nodes` input namespace. Just as for `FolderData` nodes, its contents
are copied to the working directory of the calculation job.

For `RemoteData` nodes with large files, copying the files may be
expensive. For this use case, the `metadata.options.use_symlinks` input
can be set to `True` in which case the contents are symlinked instead
of copied.