Docs: Update the styling to custom theme
The `sphinx-book-theme` is replaced with `pydata-sphinx-theme`. This
allows to change the pygments scheme for code highlighting. Additional
custom changes are made in the `docs/source/_static/custom.css` file.

The images are all moved to `docs/source/_static`. A light version is
added of the logo which is used when the dark theme is selected.