Devops: Migrate to `ruff` and cleanup pre-commit config (#52)
The `ruff` tool is adopted in the pre-commit. This completely replaces
the hooks `pylint`, `yapf`, `pydocstyle` and `isort`. The encoding
pragma is removed since this project is Python 3 only and all files are
utf-8 encoded, which is the default assumed by the Python interpreter.

For formatting, `ruff` implements the same style as `black`. The only
configuration options that are changed are to set the line length to 120
and to keep single quotes, instead of double quotes.