`ShellJob`: Automatically serialize string for `arguments`
The `arguments` input requires a `List`. It defined the `to_aiida_type`
serializer to automatically convert plain `list`s. Here, the serializer
is updated to also accept a string, which is automatically converted to
a list using `shlex.split`. This brings the interface to par with that
of the `launch_shell_job` function.