`launch_shell_job`: Add option to keep skip resolving of `command`
By default, `launch_shell_job` would resolve the command to the absolute
filepath of the corresponding executable. This would serve two purposes:
checking the command exists and increasing the reproducibility. The
relative command name may be changed on the remote over time to point to
another executable. By using the absolute filepath this probability is
reduced, but of course not fully avoided since the file at the absolute
path can still be changed on the remote.