Set default localhost scratch to `$TMP/aiida_shell_scratch`
Before, the scratch would be just `$TMP`. This would result in a the
sharded work directories of calculation jobs to be directly in the
temporary directory of the filesystem and it would be difficult to clean
it up. By putting it in a subdirectory that is clearly recognizable,
deleting stale working directories becomes a lot easier.