`prepare_computer`: Check whether `default_mpiprocs_per_machine` is set
If the `localhost` is created before `prepare_computer` is ever called,
it is possible that the property `default_mpiprocs_per_machine` is not
set. This would result in the `ShellJob` validation failing if the
scheduler type has a job resource class that is a subclass of
`aiida.schedulers.datastructures.NodeNumberJobResource`.

The `prepare_computer` function now checks that if the `localhost`
computer already exists that the `default_mpiprocs_per_machine` property
is defined if the scheduler has the relevant job resource class. If that
is the case, the default is set to 1.