`ShellJob`: Fix `RemoteData` inputs shadowing job's own input files
When a `ShellJob` would get a `RemoteData` node as input that was itself
the `remote_folder` of a previously completed `ShellJob`, the contents
of the `RemoteData` would override the input scripts of the new job
itself.

The reason is because AiiDA's engine would first create the job's input
scripts in the working directory and only then copy over the files of
the `RemoteData`. Since that folder contains many of the same files,
such as the submission script, the inputs of the new job would be
overwritten completely.

AiiDA v2.6 introduced a new feature that allows a `CalcJob` plugin to
define the order in which these files are copied over. This is leveraged
to first copy the `RemoteData` node files and copy the input files of
the job itself last.