Devops: Make use of the improved `pytest` fixtures in `aiida-core`
The new fixtures allow to configure a profile with `aiida_profile` with
a storage backend that doesn't require PostgreSQL but uses the
`core.sqlite_dos` plugin instead. The fixture by default also doesn't
configure to use RabbitMQ as the broker, however, the tests in
`aiida-shell` submit to the daemon which requires a broker. Therefore
`aiida_profile` is overridden to configure RabbitMQ.