`ShellJob`: Add `metadata.options.output_filename` to retrieve list
The `metadata.options.output_filename` is defined by the `CalcJob` base
class, and so is inherited by the `ShellJob`. The `ShellParser` plugin
actually correctly deals with this case, however, the `ShellJob` only
partially so. It did actually redirect stdout to this custom output
file, however, it did not instruct for the custom output file to be
retrieved, causing the `ERROR_OUTPUT_STDOUT_MISSING` to be returned.

The bug is fixed by replacing the default stdout filename in the
`retrieve_temporary_list` with the custom `output_filename`.