`launch_shell_job`: Move `computer` to top-level of `metadata`
The `metadata` passed to the `launch_shell_job` is just supposed to
expose the same input of the `ShellJob` and essentially forward it.
However, the `computer` input was expected to be passed nested inside
the `options` namespace, even though it is defined as a top-level
metadata input. This makes the inputs for `launch_shell_job` dfifferent
from the `ShellJob` causing unnecessary confusion.

The `computer` input should now be specified as a top-level metadata
input for the `launch_shell_job` wrapper as well and passing it nested
inside the `metadata.options` is deprecated.