`prepare_code`: Quote command when passing to `which` in order to resolve (#99)
The `command` was passed to `which` without being quoted. If the command
consists of multiple commands, which is not supported, each word would be
interpreted by `which` as a separate command, as it takes a number of commands
to resolved. This would result in a misleading error message. By quoting the
command, a multi command will correctly raise an error message for the user.